+++ datafusion/bio-format-fasta/src/storage.rs
@@ -2,7 +2,7 @@ use async_compression::tokio::bufread::GzipDecoder;
 use bytes::Bytes;
 use datafusion_bio_format_core::object_storage::{
     CompressionType, ObjectStorageOptions, get_compression_type, get_remote_stream,
-    get_remote_stream_bgzf_async, get_remote_stream_gz_async,
+    get_remote_stream_bgzf_async, get_remote_stream_gz_async, gzip_multi_member_decoder,
 };
 use futures_util::stream::BoxStream;
 use futures_util::{StreamExt, stream};
@@ -137,7 +137,7 @@ pub async fn get_local_fasta_gz_reader(
     tokio::fs::File::open(file_path)
         .await
         .map(tokio::io::BufReader::new)
-        .map(GzipDecoder::new)
+        .map(gzip_multi_member_decoder)
         .map(tokio::io::BufReader::new)
         .map(fasta::r#async::io::Reader::new)
 }