+++ datafusion/bio-format-vcf/src/storage.rs
@@ -6,6 +6,7 @@ use datafusion::arrow::datatypes::SchemaRef;
 use datafusion_bio_format_core::object_storage::{
     CompressionType, ObjectStorageOptions, StorageType, get_compression_type, get_remote_stream,
     get_remote_stream_bgzf_async, get_remote_stream_gz_async, get_storage_type,
+    gzip_multi_member_decoder,
 };
 use datafusion_bio_format_core::record_filter::RecordFieldAccessor;
 use futures::stream::BoxStream;
@@ -144,7 +145,7 @@ pub async fn get_local_vcf_gz_reader(
     tokio::fs::File::open(file_path)
         .await
         .map(tokio::io::BufReader::new)
-        .map(GzipDecoder::new)
+        .map(gzip_multi_member_decoder)
         .map(tokio::io::BufReader::new)
         .map(vcf::r#async::io::Reader::new)
 }